test: consolidate Chat coverage boundaries
