+++ server/chatmutation/operation_owner_test.go
@@ -6,18 +6,17 @@ import (
 	"testing"
 	"time"
 
+	"core/server/session"
 	"core/server/sessionruntime"
 	chatpb "core/shared/protoapi/gen/kent/api/chat"
+	chatsettingspb "core/shared/protoapi/gen/kent/api/chat_settings"
 	sessionlaunchpb "core/shared/protoapi/gen/kent/api/session_launch"
 	"core/shared/runtimeids"
 	"core/shared/serverapi"
 )
 
-func TestOperationOwnerShutdownRejectsCancelsAndJoinsInFlightOperation(t *testing.T) {
-	owner, err := NewOperationOwner(time.Second)
-	if err != nil {
-		t.Fatalf("NewOperationOwner: %v", err)
-	}
+func TestOperationOwnerShutdownRejectsCancelsAndJoins(t *testing.T) {
+	owner := newTestOperationOwner()
 	started := make(chan struct{})
 	canceled := make(chan error, 1)
 	allowExit := make(chan struct{})
@@ -43,26 +42,21 @@ func TestOperationOwnerShutdownRejectsCancelsAndJoinsInFlightOperation(t *testin
 		t.Fatalf("Close returned before operation joined: %v", err)
 	default:
 	}
-	if _, err := owner.Start(t.Context(), func(OperationScope) error {
-		return nil
-	}); !errors.Is(err, ErrOperationOwnerClosed) {
+	if _, err := owner.Start(t.Context(), func(OperationScope) error { return nil }); !errors.Is(err, ErrOperationOwnerClosed) {
 		t.Fatalf("Start after shutdown error = %v, want closed owner", err)
 	}
-
 	close(allowExit)
 	if err := <-closed; err != nil {
 		t.Fatalf("Close: %v", err)
 	}
-	ctx, cancel := context.WithTimeout(t.Context(), time.Second)
-	defer cancel()
-	if err := operation.Await(ctx); !errors.Is(err, ErrOperationOwnerClosed) {
+	if err := operation.Await(t.Context()); !errors.Is(err, ErrOperationOwnerClosed) {
 		t.Fatalf("operation result = %v, want shutdown cancellation", err)
 	}
 }
 
-func TestOperationOwnerBoundsAttachmentFinalizationDuringShutdown(t *testing.T) {
-	const finalizationTimeout = 20 * time.Millisecond
-	owner, err := NewOperationOwner(finalizationTimeout)
+func TestOperationOwnerBoundsAttachmentFinalization(t *testing.T) {
+	const timeout = 20 * time.Millisecond
+	owner, err := NewOperationOwner(timeout)
 	if err != nil {
 		t.Fatalf("NewOperationOwner: %v", err)
 	}
@@ -78,112 +72,103 @@ func TestOperationOwnerBoundsAttachmentFinalizationDuringShutdown(t *testing.T) 
 		t.Fatalf("Start: %v", err)
 	}
 	<-finalizing
-
-	startedAt := time.Now()
-	closeErr := owner.Close()
-	if !errors.Is(closeErr, context.DeadlineExceeded) {
-		t.Fatalf("Close error = %v, want bounded finalization timeout", closeErr)
+	if err := owner.Close(); !errors.Is(err, context.DeadlineExceeded) {
+		t.Fatalf("Close error = %v, want bounded finalization timeout", err)
 	}
-	if elapsed := time.Since(startedAt); elapsed < finalizationTimeout {
-		t.Fatalf("Close elapsed = %v, want at least %v", elapsed, finalizationTimeout)
+	if err := operation.Await(t.Context()); !errors.Is(err, context.DeadlineExceeded) {
+		t.Fatalf("operation result = %v, want deadline exceeded", err)
 	}
-	ctx, cancel := context.WithTimeout(t.Context(), time.Second)
-	defer cancel()
-	if err := operation.Await(ctx); !errors.Is(err, context.DeadlineExceeded) {
-		t.Fatalf("operation finalization error = %v, want deadline exceeded", err)
-	}
-}
-
-type lifecycleContextRecorder struct {
-	operation context.Context
-	caller    context.Context
-	marker    *struct{}
-	stages    map[string]struct{}
 }
 
 type lifecycleContextKey struct{}
 
-func (r *lifecycleContextRecorder) operationStage(t *testing.T, stage string, ctx context.Context) {
-	t.Helper()
-	if ctx == r.caller {
-		t.Fatalf("%s remained owned by the caller", stage)
-	}
-	if ctx.Value(lifecycleContextKey{}) != r.marker {
-		t.Fatalf("%s lost the operation context values", stage)
+type lifecycleRecorder struct {
+	t         *testing.T
+	caller    context.Context
+	marker    *struct{}
+	operation context.Context
+	stages    map[string]struct{}
+}
+
+func (r *lifecycleRecorder) operationStage(stage string, ctx context.Context) {
+	r.t.Helper()
+	if ctx == r.caller || ctx.Value(lifecycleContextKey{}) != r.marker {
+		r.t.Fatalf("%s did not receive the server-owned operation context", stage)
 	}
 	if r.operation == nil {
 		r.operation = ctx
 	} else if ctx != r.operation {
-		t.Fatalf("%s received a different operation context", stage)
-	}
-	if r.stages == nil {
-		r.stages = make(map[string]struct{})
+		r.t.Fatalf("%s received a different operation context", stage)
 	}
 	r.stages[stage] = struct{}{}
 }
 
-func (r *lifecycleContextRecorder) finalizationStage(t *testing.T, ctx context.Context) {
-	t.Helper()
-	if _, ok := ctx.Deadline(); !ok {
-		t.Fatal("attachment finalization was not bounded")
-	}
-	if err := context.Cause(ctx); err != nil {
-		t.Fatalf("attachment finalization inherited cancellation: %v", err)
-	}
-	if ctx.Value(lifecycleContextKey{}) != r.marker {
-		t.Fatal("attachment finalization lost the operation context values")
-	}
-	if r.stages == nil {
-		r.stages = make(map[string]struct{})
+func (r *lifecycleRecorder) finalizationStage(ctx context.Context) {
+	r.t.Helper()
+	if _, bounded := ctx.Deadline(); !bounded ||
+		context.Cause(ctx) != nil ||
+		ctx.Value(lifecycleContextKey{}) != r.marker {
+		r.t.Fatal("attachment finalization did not receive the bounded operation context")
 	}
 	r.stages["attachment finalization"] = struct{}{}
 }
 
-type lifecycleSessionCreation struct {
-	t         *testing.T
-	recorder  *lifecycleContextRecorder
+type lifecyclePersistedSessions struct {
+	recorder  *lifecycleRecorder
 	sessionID runtimeids.SessionID
 }
 
-func (s lifecycleSessionCreation) PlanSession(
+func (s lifecyclePersistedSessions) ResolvePersistedSession(
+	ctx context.Context,
+	_ string,
+) (session.PersistedSessionRecord, error) {
+	s.recorder.operationStage("target resolution", ctx)
+	return session.PersistedSessionRecord{
+		SessionDir: "/tmp/" + s.sessionID.String(),
+		Meta:       &session.Meta{SessionID: s.sessionID.String()},
+	}, nil
+}
+
+type lifecycleCreation struct {
+	recorder  *lifecycleRecorder
+	sessionID runtimeids.SessionID
+}
+
+func (s lifecycleCreation) PlanSession(
 	ctx context.Context,
 	_ *sessionlaunchpb.SessionPlanRequest,
 ) (*sessionlaunchpb.SessionPlanSuccess, error) {
-	s.recorder.operationStage(s.t, "Session creation", ctx)
+	s.recorder.operationStage("Session creation", ctx)
 	return &sessionlaunchpb.SessionPlanSuccess{
 		Plan: &sessionlaunchpb.SessionPlan{SessionId: s.sessionID.String()},
 	}, nil
 }
 
-type lifecycleRuntimePlanner struct {
-	t          *testing.T
-	recorder   *lifecycleContextRecorder
+type lifecyclePlanner struct {
+	recorder   *lifecycleRecorder
 	attachment RuntimeAttachment
 }
 
-func (p lifecycleRuntimePlanner) Open(
+func (p lifecyclePlanner) Open(
 	ctx context.Context,
 	_ runtimeids.SessionID,
 ) (RuntimeAttachment, error) {
-	p.recorder.operationStage(p.t, "Runtime opening", ctx)
+	p.recorder.operationStage("Runtime opening", ctx)
 	return p.attachment, nil
 }
 
 type lifecycleAdmission struct {
-	t           *testing.T
-	recorder    *lifecycleContextRecorder
-	queueItemID runtimeids.QueueItemID
+	recorder *lifecycleRecorder
+	result   serverapi.ChatInputAdmissionResult
+	err      error
 }
 
 func (a lifecycleAdmission) AdmitChatUserTurn(
 	ctx context.Context,
 	_ serverapi.RuntimeSubmitUserTurnRequest,
 ) (serverapi.ChatInputAdmissionResult, error) {
-	a.recorder.operationStage(a.t, "admission", ctx)
-	return serverapi.ChatInputAdmissionResult{
-		QueueItemID: a.queueItemID,
-		Accepted:    true,
-	}, nil
+	a.recorder.operationStage("admission", ctx)
+	return a.result, a.err
 }
 
 func (lifecycleAdmission) AdmitChatQueuedUserInput(
@@ -201,82 +186,151 @@ func (lifecycleAdmission) AdmitManualCompaction(
 }
 
 type lifecycleAttachment struct {
-	t         *testing.T
-	recorder  *lifecycleContextRecorder
+	recorder  *lifecycleRecorder
 	sessionID runtimeids.SessionID
+	policy    sessionruntime.RuntimeReleasePolicy
 }
 
-func (a lifecycleAttachment) SessionID() runtimeids.SessionID {
+func (a *lifecycleAttachment) SessionID() runtimeids.SessionID {
 	return a.sessionID
 }
 
-func (a lifecycleAttachment) Release(
+func (a *lifecycleAttachment) Release(
 	ctx context.Context,
 	policy sessionruntime.RuntimeReleasePolicy,
 ) error {
-	if policy != sessionruntime.RuntimeReleaseDetach {
-		a.t.Fatalf("release policy = %v, want detach", policy)
-	}
-	a.recorder.finalizationStage(a.t, ctx)
+	a.policy = policy
+	a.recorder.finalizationStage(ctx)
 	return nil
 }
 
-func TestServiceCarriesOneOperationContextThroughFiveStageLifecycle(t *testing.T) {
-	owner := newTestOperationOwner()
-	t.Cleanup(func() {
-		if err := owner.Close(); err != nil {
-			t.Errorf("close operation owner: %v", err)
-		}
-	})
-	sessionID := runtimeids.NewSessionID()
-	queueItemID := runtimeids.NewQueueItemID()
-	marker := &struct{}{}
-	callerCtx := context.WithValue(t.Context(), lifecycleContextKey{}, marker)
-	recorder := &lifecycleContextRecorder{caller: callerCtx, marker: marker}
-	creation := lifecycleSessionCreation{
-		t:         t,
-		recorder:  recorder,
-		sessionID: sessionID,
-	}
-	targets := NewTargetResolver(
-		nil,
-		func(ctx context.Context, _, _ string) (SessionCreationService, error) {
-			recorder.operationStage(t, "target resolution", ctx)
-			return creation, nil
+func TestServiceCarriesOneOperationContextAcrossChatLifecycle(t *testing.T) {
+	tests := []struct {
+		name         string
+		newChat      bool
+		admission    serverapi.ChatInputAdmissionResult
+		admissionErr error
+		wantPolicy   sessionruntime.RuntimeReleasePolicy
+		wantStages   []string
+	}{
+		{
+			name:    "New Chat accepted with typed prompt-history diagnostic",
+			newChat: true,
+			admission: serverapi.ChatInputAdmissionResult{
+				QueueItemID:          runtimeids.NewQueueItemID(),
+				Accepted:             true,
+				PromptHistoryFailure: errors.New("prompt history unavailable"),
+			},
+			wantPolicy: sessionruntime.RuntimeReleaseDetach,
+			wantStages: []string{
+				"target resolution",
+				"Session creation",
+				"Runtime opening",
+				"admission",
+				"attachment finalization",
+			},
 		},
-	)
-	attachment := lifecycleAttachment{t: t, recorder: recorder, sessionID: sessionID}
-	service := NewService(
-		owner,
-		targets,
-		lifecycleRuntimePlanner{t: t, recorder: recorder, attachment: attachment},
-		lifecycleAdmission{t: t, recorder: recorder, queueItemID: queueItemID},
-	)
+		{
+			name:         "existing Session rejected before acceptance",
+			admissionErr: &serverapi.PendingWorkCapacityError{},
+			wantPolicy:   sessionruntime.RuntimeReleaseCloseIfIdle,
+			wantStages: []string{
+				"target resolution",
+				"Runtime opening",
+				"admission",
+				"attachment finalization",
+			},
+		},
+	}
+	for _, test := range tests {
+		t.Run(test.name, func(t *testing.T) {
+			owner := newTestOperationOwner()
+			t.Cleanup(func() { _ = owner.Close() })
+			sessionID := runtimeids.NewSessionID()
+			marker := &struct{}{}
+			caller := context.WithValue(t.Context(), lifecycleContextKey{}, marker)
+			recorder := &lifecycleRecorder{
+				t:      t,
+				caller: caller,
+				marker: marker,
+				stages: make(map[string]struct{}),
+			}
+			targets := NewTargetResolver(
+				lifecyclePersistedSessions{recorder: recorder, sessionID: sessionID},
+				func(ctx context.Context, _, _ string) (SessionCreationService, error) {
+					recorder.operationStage("target resolution", ctx)
+					return lifecycleCreation{recorder: recorder, sessionID: sessionID}, nil
+				},
+			)
+			attachment := &lifecycleAttachment{recorder: recorder, sessionID: sessionID}
+			service := NewService(
+				owner,
+				targets,
+				lifecyclePlanner{recorder: recorder, attachment: attachment},
+				lifecycleAdmission{
+					recorder: recorder,
+					result:   test.admission,
+					err:      test.admissionErr,
+				},
+			)
+			target := &chatpb.ChatTarget{
+				Target: &chatpb.ChatTarget_Session{
+					Session: &chatpb.ExistingSessionTarget{SessionId: sessionID.String()},
+				},
+			}
+			if test.newChat {
+				target.Target = &chatpb.ChatTarget_NewChat{NewChat: validNewChatTarget()}
+			}
 
-	result, err := service.Steer(callerCtx, &chatpb.SteerRequest{
-		Target: &chatpb.ChatTarget{
-			Target: &chatpb.ChatTarget_NewChat{NewChat: validSteerNewChatTarget()},
-		},
-		Activation: &chatpb.Activation{
-			Input: &chatpb.Activation_Text{Text: "continue"},
-		},
-	})
+			result, err := service.Steer(caller, &chatpb.SteerRequest{
+				Target: target,
+				Activation: &chatpb.Activation{
+					Input: &chatpb.Activation_Text{Text: "continue"},
+				},
+			})
+			if err != nil {
+				t.Fatalf("Steer: %v", err)
+			}
+			if result.GetSession().GetSessionId() != sessionID.String() {
+				t.Fatalf("Steer Session = %+v, want %s", result.GetSession(), sessionID)
+			}
+			if test.admission.Accepted {
+				if result.GetAccepted().GetDiagnostic().GetPromptHistoryFailure() == nil {
+					t.Fatalf("Steer result = %+v, want typed prompt-history diagnostic", result)
+				}
+			} else if result.GetNotAccepted().GetPendingWorkCapacity() == nil {
+				t.Fatalf("Steer result = %+v, want capacity rejection", result)
+			}
+			if attachment.policy != test.wantPolicy {
+				t.Fatalf("release policy = %v, want %v", attachment.policy, test.wantPolicy)
+			}
+			for _, stage := range test.wantStages {
+				if _, observed := recorder.stages[stage]; !observed {
+					t.Errorf("%s did not receive the operation context", stage)
+				}
+			}
+		})
+	}
+}
+
+func newTestOperationOwner() *OperationOwner {
+	owner, err := NewOperationOwner(time.Second)
 	if err != nil {
-		t.Fatalf("Steer: %v", err)
+		panic(err)
 	}
-	if result.GetSession().GetSessionId() != sessionID.String() ||
-		result.GetAccepted().GetQueueItem().GetId() != queueItemID.String() {
-		t.Fatalf("Steer result = %+v", result)
-	}
-	for _, stage := range []string{
-		"target resolution",
-		"Session creation",
-		"Runtime opening",
-		"admission",
-		"attachment finalization",
-	} {
-		if _, observed := recorder.stages[stage]; !observed {
-			t.Errorf("%s did not receive the operation context", stage)
-		}
+	return owner
+}
+
+func validNewChatTarget() *chatpb.NewChatTarget {
+	questions, autoCompaction := true, true
+	return &chatpb.NewChatTarget{
+		ProjectId:   "project-1",
+		WorkspaceId: "workspace-1",
+		InitialSettings: &chatpb.InitialChatSettings{
+			AgentRole:             "default",
+			Supervisor:            chatsettingspb.SupervisorValue_SUPERVISOR_VALUE_OFF,
+			QuestionsEnabled:      &questions,
+			AutoCompactionEnabled: &autoCompaction,
+		},
 	}
 }
