+++ server/transport/gateway_test.go
@@ -19,9 +19,11 @@ import (
 	"core/internal/testharness/testsetup"
 	"core/server/auth"
 	serverbootstrap "core/server/bootstrap"
+	"core/server/chatmutation"
 	"core/server/core"
 	"core/server/metadata"
 	"core/server/session"
+	"core/server/sessionruntime"
 	shelltool "core/server/tools/shell"
 	"core/shared/apicontract"
 	remoteclient "core/shared/client"
@@ -29,6 +31,7 @@ import (
 	"core/shared/llmerrors"
 	"core/shared/protoapi"
 	authpb "core/shared/protoapi/gen/kent/api/auth"
+	chatpb "core/shared/protoapi/gen/kent/api/chat"
 	connectionpb "core/shared/protoapi/gen/kent/api/connection"
 	projectpb "core/shared/protoapi/gen/kent/api/project"
 	serverpb "core/shared/protoapi/gen/kent/api/server"
@@ -43,6 +46,184 @@ import (
 
 	"google.golang.org/protobuf/types/known/emptypb"
 )
+
+type gatewayChatLifecycleDependencies struct {
+	GatewayDependencies
+	chat apicontract.ChatMutationService
+}
+
+func (d *gatewayChatLifecycleDependencies) ChatMutationClient() apicontract.ChatMutationService {
+	return d.chat
+}
+
+type gatewayChatLifecycleResolver struct {
+	started chan context.Context
+	proceed chan struct{}
+	target  chatmutation.ResolvedTarget
+}
+
+func (r *gatewayChatLifecycleResolver) Resolve(
+	ctx context.Context,
+	_ chatmutation.TargetResolutionRequest,
+) (chatmutation.ResolvedTarget, error) {
+	r.started <- ctx
+	select {
+	case <-r.proceed:
+		return r.target, nil
+	case <-ctx.Done():
+		return chatmutation.ResolvedTarget{}, context.Cause(ctx)
+	}
+}
+
+type gatewayChatLifecyclePlanner struct {
+	attachment chatmutation.RuntimeAttachment
+}
+
+func (p gatewayChatLifecyclePlanner) Open(
+	context.Context,
+	runtimeids.SessionID,
+) (chatmutation.RuntimeAttachment, error) {
+	return p.attachment, nil
+}
+
+type gatewayChatLifecycleAttachment struct {
+	sessionID runtimeids.SessionID
+	released  chan sessionruntime.RuntimeReleasePolicy
+}
+
+func (a *gatewayChatLifecycleAttachment) SessionID() runtimeids.SessionID {
+	return a.sessionID
+}
+
+func (a *gatewayChatLifecycleAttachment) Release(
+	_ context.Context,
+	policy sessionruntime.RuntimeReleasePolicy,
+) error {
+	a.released <- policy
+	return nil
+}
+
+type gatewayChatLifecycleAdmission struct {
+	queueItemID runtimeids.QueueItemID
+}
+
+func (a gatewayChatLifecycleAdmission) AdmitChatUserTurn(
+	context.Context,
+	serverapi.RuntimeSubmitUserTurnRequest,
+) (serverapi.ChatInputAdmissionResult, error) {
+	return serverapi.ChatInputAdmissionResult{
+		QueueItemID: a.queueItemID,
+		Accepted:    true,
+	}, nil
+}
+
+func (gatewayChatLifecycleAdmission) AdmitChatQueuedUserInput(
+	context.Context,
+	serverapi.RuntimeSubmitUserTurnRequest,
+) (serverapi.ChatInputAdmissionResult, error) {
+	panic("unexpected Queue admission")
+}
+
+func (gatewayChatLifecycleAdmission) AdmitManualCompaction(
+	context.Context,
+	serverapi.RuntimeCompactContextRequest,
+) (bool, error) {
+	panic("unexpected compaction admission")
+}
+
+func TestGatewayDisconnectStopsDeliveryWithoutCancelingChatOperation(t *testing.T) {
+	appCore, _ := newGatewayTestCore(t, true, true)
+	t.Cleanup(func() { _ = appCore.Close() })
+	store := createGatewayAuthoritativeSession(t, appCore)
+	sessionID, err := runtimeids.ParseSessionID(store.Meta().SessionID)
+	if err != nil {
+		t.Fatalf("parse Session ID: %v", err)
+	}
+	owner, err := chatmutation.NewOperationOwner(time.Second)
+	if err != nil {
+		t.Fatalf("NewOperationOwner: %v", err)
+	}
+	t.Cleanup(func() { _ = owner.Close() })
+	resolver := &gatewayChatLifecycleResolver{
+		started: make(chan context.Context, 1),
+		proceed: make(chan struct{}),
+		target:  chatmutation.ResolvedTarget{SessionID: sessionID},
+	}
+	attachment := &gatewayChatLifecycleAttachment{
+		sessionID: sessionID,
+		released:  make(chan sessionruntime.RuntimeReleasePolicy, 1),
+	}
+	service := chatmutation.NewService(
+		owner,
+		resolver,
+		gatewayChatLifecyclePlanner{attachment: attachment},
+		gatewayChatLifecycleAdmission{queueItemID: runtimeids.NewQueueItemID()},
+	)
+	gateway, err := NewGateway(
+		&gatewayChatLifecycleDependencies{GatewayDependencies: appCore, chat: service},
+		gatewayTestIdentity(),
+	)
+	if err != nil {
+		t.Fatalf("NewGateway: %v", err)
+	}
+	server := httptest.NewServer(gateway.Handler())
+	t.Cleanup(server.Close)
+	conn := dialGateway(t, server)
+	handshakeGateway(t, conn)
+
+	method := chatpb.File_kent_api_chat_chat_proto.Services().
+		ByName("ChatService").
+		Methods().
+		ByName("Steer")
+	operation, err := protoapi.OperationFromDescriptor(method)
+	if err != nil {
+		t.Fatalf("resolve Chat operation: %v", err)
+	}
+	payload, err := protoapi.Marshal(&chatpb.SteerRequest{
+		Target: &chatpb.ChatTarget{Target: &chatpb.ChatTarget_Session{
+			Session: &chatpb.ExistingSessionTarget{SessionId: sessionID.String()},
+		}},
+		Activation: &chatpb.Activation{
+			Input: &chatpb.Activation_Text{Text: "continue"},
+		},
+	})
+	if err != nil {
+		t.Fatalf("marshal Chat request: %v", err)
+	}
+	correlation := "disconnect-chat"
+	envelope, err := protoapi.EncodeEnvelope(&sharedpb.Envelope{
+		Frame: &sharedpb.Envelope_Call{Call: &sharedpb.Call{
+			Operation:   operation.Name,
+			Correlation: &correlation,
+			Payload:     payload,
+		}},
+	})
+	if err != nil {
+		t.Fatalf("encode Chat request: %v", err)
+	}
+	if err := websocket.Message.Send(conn, envelope); err != nil {
+		t.Fatalf("send Chat request: %v", err)
+	}
+
+	operationCtx := <-resolver.started
+	if err := conn.Close(); err != nil {
+		t.Fatalf("close caller connection: %v", err)
+	}
+	select {
+	case <-operationCtx.Done():
+		t.Fatalf("Gateway disconnect canceled Chat operation: %v", context.Cause(operationCtx))
+	default:
+	}
+	close(resolver.proceed)
+	select {
+	case policy := <-attachment.released:
+		if policy != sessionruntime.RuntimeReleaseDetach {
+			t.Fatalf("release policy = %v, want detach", policy)
+		}
+	case <-time.After(time.Second):
+		t.Fatal("Chat operation did not finish after Gateway disconnect")
+	}
+}
 
 func gatewaySessionExecutionTarget(t *testing.T, conn *websocket.Conn, requestID, sessionID string) clientui.SessionExecutionTarget {
 	t.Helper()
